Backend: protegiendo todas las rutas de proyectos. Requiere Auth

+++ uptask-backend/src/controllers/ProjectController.ts
@@ -19,7 +19,9 @@ export class ProjectController {
     static getAllProjects = async (req : Request, res : Response) => {
         try {
             const projects = await Project.find({
-
+                $or: [
+                    { manager: { $in: req.user.id} }
+                ]
             })
 
             res.json(projects)
@@ -38,6 +40,11 @@ export class ProjectController {
                 return res.status(404).json({ error: error.message })
             }
 
+            if(project.manager.toString() !== req.user.id.toString()) {
+                const error = new Error('Acción No Válida')
+                return res.status(401).json({ error: error.message })
+            }
+
             res.json(project)
         } catch (error) {
             console.log(error.message)
@@ -48,11 +55,15 @@ export class ProjectController {
         try {
             const { id } = req.params
             const project = await Project.findById(id)
-
             
             if(!project) { 
                 const error = new Error('Project not found')
                 return res.status(404).json({ error: error.message })
+            }
+
+            if(project.manager.toString() !== req.user.id.toString()) {
+                const error = new Error('Acción No Válida')
+                return res.status(401).json({ error: error.message })
             }
 
             project.projectName = req.body.projectName
@@ -70,14 +81,18 @@ export class ProjectController {
         try {
             const { id } = req.params
             const project = await Project.findById(id)
-
-            await project.deleteOne()
-
+            
             if(!project) { 
                 const error = new Error('Project not found')
                 return res.status(404).json({ error: error.message })
             }
 
+            if(project.manager.toString() !== req.user.id.toString()) {
+                const error = new Error('Acción No Válida')
+                return res.status(401).json({ error: error.message })
+            }
+            
+            await project.deleteOne()
             res.send('Project Deleted Successfully')
         } catch (error) {
             console.error(error.message)
